+++ conflict.h
@@ -2,5 +2,6 @@
 #define CONFLICT_H_
 
 int* getConflictsPerQueen(int N, int* board);
+int computeConflictScore(int N, int *board);
 
 #endif
