+++ test.h
@@ -6,6 +6,7 @@ void fillRandom (int N, int* board);
 void testMutateTrulyRandom();
 void testCrossoverRandomSplit();
 void testGetConflictsPerQueen();
+void testComputeConflictScore();
 void testMax();
 void testGetIndex();
 void testMutate();
